Feat(tag-category): create and update tag category with category, throw error if not found when delete

+++ src/tag-categories/tag-categories.service.ts
@@ -1,21 +1,37 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TagCategory } from './tag-category.entity';
 import { Repository } from 'typeorm';
-import { Tag } from 'src/tags/tag.entity';
 import { UpdateTagCategoryDto } from './dto/udpate-tag.dto';
-import { CreateTagCategoryDto } from './dto/create-tag.dto';
+import { CreateTagCategoryDto } from './dto/create-tag-category.dto';
+import { Category } from 'src/categories/category.entity';
 
 @Injectable()
 export class TagCategoriesService {
   constructor(
     @InjectRepository(TagCategory)
-    private tagCategoryRepository: Repository<Tag>,
+    private tagCategoryRepository: Repository<TagCategory>,
+    @InjectRepository(Category)
+    private categoriesRepository: Repository<Category>,
   ) {}
 
-  createTagCategory(createTagCategoryDto: CreateTagCategoryDto) {
-    const newTagCategory =
-      this.tagCategoryRepository.create(createTagCategoryDto);
+  async createTagCategory(createTagCategoryDto: CreateTagCategoryDto) {
+    const { categoryId, ...tagCategory } = createTagCategoryDto;
+
+    const newTagCategory = this.tagCategoryRepository.create(tagCategory);
+
+    if (categoryId) {
+      const category = await this.categoriesRepository.findOneBy({
+        id: categoryId,
+      });
+
+      if (!category) {
+        throw new NotFoundException(`Category with id ${categoryId} not found`);
+      } else {
+        newTagCategory.category = category;
+      }
+    }
+
     return this.tagCategoryRepository.save(newTagCategory);
   }
 
@@ -25,12 +41,31 @@ export class TagCategoriesService {
 
   async deleteTagCategory(id: string) {
     const tagCategory = await this.tagCategoryRepository.findOneBy({ id });
-    return this.tagCategoryRepository.remove(tagCategory);
+
+    if (!tagCategory) {
+      throw new NotFoundException(`Tag category with id ${id} not found`);
+    } else {
+      return this.tagCategoryRepository.remove(tagCategory);
+    }
   }
 
   async updateTagCategory(updateTagCategoryDto: UpdateTagCategoryDto) {
-    const { id, ...tagCategory } = updateTagCategoryDto;
+    const { id, categoryId, ...tagCategory } = updateTagCategoryDto;
+
+    if (categoryId) {
+      const category = await this.categoriesRepository.findOneBy({
+        id: categoryId,
+      });
+
+      if (!category) {
+        throw new NotFoundException(`Category with id ${categoryId} not found`);
+      } else {
+        tagCategory.category = category;
+      }
+    }
+
     await this.tagCategoryRepository.update({ id }, tagCategory);
+
     return this.tagCategoryRepository.findOneBy({ id });
   }
 }
